requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==5.1.7
django-allauth==0.50.0
django-environ==0.12.0
django-sendgrid-v5==1.3.0
ecdsa==0.19.1
gunicorn==20.1.0
idna==3.10
MarkupSafe==3.0.2
oauthlib==3.2.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.1
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
sendgrid==6.12.4
setuptools==80.8.0
six==1.17.0
sqlparse==0.5.3
stripe==12.1.0
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
whitenoise==6.9.0
